+++ Utilities/Mines.cs
@@ -36,22 +36,22 @@ namespace SeedFinding
     {
         public static List<Geode> allGeodes = new() { Geode.Geode, Geode.FrozenGeode, Geode.MagmaGeode, Geode.OmniGeode, Geode.Trove, Geode.Coconut, Geode.MysteryBox, Geode.GoldenMysteryBox };
 
-        public static List<int> CheckStone1_5(int gameId, int floor, int x, int y, bool ladder=false, bool geologist = false, bool excavator=false)
+        public static List<int> CheckStone1_5(long gameId, int floor, int x, int y, bool ladder=false, bool geologist = false, bool excavator=false)
         {
             return CheckStone1_5(x*1000 + y + floor + gameId / 2, ladder,geologist,excavator,floor);
         }
 
-        public static List<int> CheckStone1_6(int day, uint gameId, int floor, int x, int y, bool ladder = false, bool geologist = false, bool excavator = false)
+        public static List<int> CheckStone1_6(int day, long gameId, int floor, int x, int y, bool ladder = false, bool geologist = false, bool excavator = false)
         {
 
             Random r = Utility.CreateDaySaveRandom(day, gameId, x * 1000, y, floor);
             return CheckStone(r, ladder, geologist, excavator,floor);
         }
 
-        public static List<int> CheckStone1_5(int seed, bool ladder = false, bool geologist = false, bool excavator = false, int floor = 0)
+        public static List<int> CheckStone1_5(long seed, bool ladder = false, bool geologist = false, bool excavator = false, int floor = 0)
         {
 
-            Random r = new Random(seed);
+            Random r = new Random((int)seed);
             return CheckStone(r, ladder, geologist, excavator, floor);
         }
 
@@ -111,7 +111,7 @@ namespace SeedFinding
             return value;
         }
 
-        public static (string,int) GetGeodeContents1_6(int gameId, int geodesCracked, Geode whichGeode, long playerId, int deepestMineLevel=0, bool qibeans = false, bool farmingMastery=false, int fishingLevel=0, bool hasCoconutHat = false, bool hasMysteryBook = false)
+        public static (string,int) GetGeodeContents1_6(long gameId, int geodesCracked, Geode whichGeode, long playerId, int deepestMineLevel=0, bool qibeans = false, bool farmingMastery=false, int fishingLevel=0, bool hasCoconutHat = false, bool hasMysteryBook = false)
         {
             Random r = Utility.CreateRandom(geodesCracked, gameId / 2, (int)playerId/2); 
             int prewarm_amount = r.Next(1, 10);
@@ -504,7 +504,7 @@ namespace SeedFinding
                 _ => ("(O)382", amount),
             };
         }
-        public static void PrintGeodeContents(int gameId, long playerId, int startingGeode, int count, List<Geode> geodeTypes, string delimiter, bool excludeOres=true, int deepestMineLevel=0, bool qibeans = false, bool printBestGeode=false, int printBestGeodeMinPrice=0, bool before1_5=false, StreamWriter stream = null)
+        public static void PrintGeodeContents(long gameId, long playerId, int startingGeode, int count, List<Geode> geodeTypes, string delimiter, bool excludeOres=true, int deepestMineLevel=0, bool qibeans = false, bool printBestGeode=false, int printBestGeodeMinPrice=0, bool before1_5=false, StreamWriter stream = null)
 
         {
             List<string> unsellables = new List<string>() { "100", "101", "103", "104", "105", "106", "108", "109", "110", "111", "112", "113", "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "330", "390" };
